+++ CHANGELOG.md
@@ -12,8 +12,11 @@ This document outlines major changes between releases.
 ### Changed
 
 ### Removed
+* RANGEHASH support (#391)
 
 ### Updated
+- github.com/nspcc-dev/neofs-sdk-go dependency to `v1.0.0-rc.19` (#391)
+- github.com/nspcc-dev/neo-go dependency to `v0.120.0` (#391)
 
 ### Upgrading from 0.17.0